pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-vector-db-benchmark"
version = "0.1.0"
description = "A research-grade framework for systematically evaluating Retrieval-Augmented Generation (RAG) systems with a focus on vector database performance, retrieval quality, and end-to-end pipeline metrics."
authors = [{ name = "mintmaint", email = "test@example.com" }]
requires-python = ">=3.9"
readme = "README.md"
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "PyYAML>=6.0",
    "psutil>=5.9.0",
    "tqdm>=4.66.0",
    "datasets>=2.14.0",
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
    "litellm>=1.30.0",
    "chromadb>=0.4.0",
    "qdrant-client>=1.7.0",
    "pinecone-client>=3.0.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.22.0",
    "pydantic>=2.0.0",
    "httpx>=0.24.0",
    "gradio>=4.0.0",
    "pypdf>=4.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "jupyter>=1.0.0"
]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-v --tb=short"
testpaths = [
    "tests",
]
python_files = ["test_*.py"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true

[[tool.mypy.overrides]]
module = [
    "yaml.*",
    "chromadb.*",
    "pinecone.*",
    "sentence_transformers.*",
    "litellm.*",
    "gradio.*",
    "datasets.*",
    "fastapi.*",
    "pypdf.*",
    "tqdm.*"
]
ignore_missing_imports = true

[tool.pyright]
include = ["src", "tests", "scripts", "api.py", "app_gui.py"]
extraPaths = ["."]
